Move bearer to utils
